requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2023.10
Django==5.0.2
django-crispy-forms==2.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
